% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_data.R
\docType{data}
\name{data_full_rho0.2}
\alias{data_full_rho0.2}
\title{Mock/simulated full data set Rho=0.2}
\format{An object of class \code{data.frame} with 1000000 rows and 14 columns.}
\usage{
data_full_rho0.2
}
\description{
Datasets containing various dependence structures between a surrogate S and
the final outcome of interest Y, to be used with the estimation routines
implemented.
}
\keyword{datasets}
