% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_data.R
\docType{data}
\name{data_rho0.8}
\alias{data_rho0.8}
\title{Mock/simulated and sampled data set Rho=0.8}
\format{An object of class \code{data.frame} with 1000 rows and 14 columns.}
\usage{
data_rho0.8
}
\description{
Datasets containing various dependence structures between a surrogate S and
the final outcome of interest Y, to be used with the estimation routines
implemented.
}
\keyword{datasets}
