% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmle3_surrogate_adapt.R
\name{tmle3_sadapt}
\alias{tmle3_sadapt}
\title{Learning of the Optimal Surrogate from the observed final outcome Y, for the
Mean Under the Optimal Individualized Rule and Average Treatment Effect}
\usage{
tmle3_sadapt(surrogate = TRUE, S = S, W = c("W1", "W2", "W3"),
  V = NULL, A = "A", Y = "Y", learners, data, old_data,
  param = "opt", training_size = 100, test_size = 20,
  mini_batch = 25, Gexploit = 0.1, Gexplore = 0.01, n_max = 1600,
  by = 200, n = nrow(data), gen_data = gen_data,
  gen_data_adapt = gen_data_adapt, ...)
}
\arguments{
\item{surrogate}{If \code{TRUE}, adaptively learns the target parameter with
 respect to the
estimated optimal surrogate.}

\item{S}{Covariates to consider for the Optimal Surrogate estimation.}

\item{W}{Baseline covariates used in the trial.}

\item{V}{Covariates the rule depends on. Not used currently.}

\item{A}{Treatment used on the trial.}

\item{Y}{Outcome node used in the trial.}

\item{learners}{\code{list} of learners used for Q,g,S and B.}

\item{data}{...}

\item{old_data}{...}

\item{param}{Target parameter. Current implementation supports Mean under the
Optimal Individualized Treatment (opt) and Average Treatment Effect (are)}

\item{training_size}{Size of the initial training set. Necessary part of
online Super Learner.}

\item{test_size}{Size of the test set. Necessary part of online Super
Learner.}

\item{mini_batch}{Size of the increase in the initial training size, added
per each iteration of the online Super Learner.}

\item{Gexploit}{Sequence t_n. Default is 0.1.}

\item{Gexplore}{Sequence e_n. Default is 0.05.}

\item{n_max}{Maximum sample size after which the trial ends.}

\item{by}{Size of the iterative sample size increase in the sequential trial.}

\item{n}{Initial sample size.}

\item{gen_data}{Data generating distribution of the process. Used for
simulations.}

\item{gen_data_adapt}{Data generating distribution of the process with
user-specifed rule. Used for simulations.}

\item{...}{Additional arguments passed to the constructor of the \code{R6}
class \code{tmle3_surrogate}.}
}
\description{
O = (W,A,S,Y)
W = Covariates
A = Treatment (binary)
S = Potential Surrogates
Y = Outcome (binary or bounded continuous)
}
