% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmle3_Spec_surrogate.R
\docType{data}
\name{tmle3_Spec_surrogate}
\alias{tmle3_Spec_surrogate}
\title{Learns the Optimal Surrogate based on the observed final outcome Y, for the
Mean Under the Optimal Individualized Rule and Average Treatment Effect
target parameters.}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
tmle3_Spec_surrogate
}
\description{
Learns the Optimal Surrogate based on the observed final outcome Y, for the
Mean Under the Optimal Individualized Rule and Average Treatment Effect
target parameters.
}
\keyword{datasets}
