% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmle3_Spec_surrogate.R
\name{tmle3_surrogate}
\alias{tmle3_surrogate}
\title{Learns the Optimal Surrogate based on the observed final outcome Y, for the
Mean Under the Optimal Individualized Rule and Average Treatment Effect
target parameters.}
\usage{
tmle3_surrogate(S, V = NULL, learners, param = "opt",
  training_size = NULL, test_size = NULL, mini_batch = NULL,
  rule_outcome = "S", opt_surrogate = "SL")
}
\arguments{
\item{S}{Covariates to consider for the Optimal Surrogate estimation.}

\item{V}{Covariates the rule depends on.}

\item{learners}{List of learners used for Q,g,S and B.}

\item{param}{Target parameter. Current implementation supports Mean under the Optimal Individualized
Treatment (opt) and Average Treatment Effect (are)}

\item{training_size}{Size of the initial training set. Necessary part of online Super Learner.}

\item{test_size}{Size of the test set. Necessary part of online Super Learner.}

\item{mini_batch}{Size of the increase in the initial training size, added per each iteration of the
online Super Learner.}

\item{opt_surrogate}{Optimal Surrogate can be based on the Super Learner prediction ("SL"), or
targeted Super Learner ("TMLE").}
}
\description{
O=(W,A,S,Y)
W=Covariates
A=Treatment (binary)
S=Potential Surrogates
Y=Outcome (binary or bounded continuous)
}
