% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmle3_Update_adapt.R
\docType{data}
\name{tmle3_Update_adapt}
\alias{tmle3_Update_adapt}
\title{Defines an update (submodel+loss function)
Allows for adding custom weights to the fluctuation step.}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
tmle3_Update_adapt
}
\description{
Current Limitations:
only updating one node
loss function and submodel are hard-coded (need to accept arguments for these)
no support for one-step (recursive TMLE)
}
\keyword{datasets}
