% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmle3_Spec_surrogate.R
\name{tmle3_surrogate}
\alias{tmle3_surrogate}
\title{Learns the Optimal Surrogate based on the observed final outcome Y, for the
Mean Under the Optimal Individualized Rule and Average Treatment Effect
target parameters.}
\usage{
tmle3_surrogate(S, V = NULL, learners, param = "opt",
<<<<<<< HEAD
  training_size = NULL, test_size = NULL, mini_batch = NULL, ...)
=======
  training_size = NULL, test_size = NULL, mini_batch = NULL,
  rule_outcome = "S", opt_surrogate = "SL")
>>>>>>> fb5d082b8a95e093206d5eb52b97326fde9d39e1
}
\arguments{
\item{S}{Covariates to consider for the Optimal Surrogate estimation.}

\item{V}{Covariates the rule depends on.}

\item{learners}{List of learners used for Q,g,S and B.}

\item{param}{Target parameter. Current implementation supports Mean under the
Optimal Individualized Treatment (opt) and Average Treatment Effect (are)}

\item{training_size}{Size of the initial training set. Necessary part of
online Super Learner.}

\item{test_size}{Size of the test set. Necessary part of online Super Learner.}

<<<<<<< HEAD
\item{mini_batch}{Size of the increase in the initial training size, added
per each iteration of the online Super Learner.}

\item{...}{Additional arguments passed to the constructor of the superclass.}
=======
\item{mini_batch}{Size of the increase in the initial training size, added per each iteration of the
online Super Learner.}

\item{rule_outcome}{If the target parameter is the Mean under the Optimal Individualized Treatment,
then the rule can be learned with respect to the surrogate ("S"), or the actual final outcome ("Y").}

\item{opt_surrogate}{Optimal Surrogate can be based on the Super Learner prediction ("SL"), or
targeted Super Learner ("TMLE"). In the case Optimal Surrogate is based on the targeted Super Learner,
epsilon from the first part of the trial is transferred and propagated adaptively to further sequential trials.}
>>>>>>> fb5d082b8a95e093206d5eb52b97326fde9d39e1
}
\description{
O=(W,A,S,Y)
W=Covariates
A=Treatment (binary)
S=Potential Surrogates
Y=Outcome (binary or bounded continuous)
}
